refactor(UI): combined project details settings in less components
