+++ ScriptBeeClient/src/app/pages/projects/project-details/settings/delete-project-dialog/delete-project-dialog.component.ts
