--- ScriptBeeClient/src/app/pages/projects/project-details/settings/project-danger-zone/project-danger-zone.component.html
@@ -1,22 +1,0 @@
-<div>
-  <h4>Danger Zone</h4>
-
-  <mat-card appearance="outlined">
-    <mat-card-header>
-      <mat-card-title>Delete Project</mat-card-title>
-      <mat-card-subtitle> Id: {{ project().id }}</mat-card-subtitle>
-      <mat-card-subtitle> Name: {{ project().name }}</mat-card-subtitle>
-    </mat-card-header>
-
-    <mat-card-content>
-      <p>Once project is deleted, all data will be removed</p>
-
-      @if (isDeleteLoading()) {
-        <app-loading-progress-bar title="Deleting project..."></app-loading-progress-bar>
-      }
-    </mat-card-content>
-    <mat-card-actions>
-      <button mat-flat-button (click)="openDialog()">DELETE</button>
-    </mat-card-actions>
-  </mat-card>
-</div>
