--- ScriptBeeClient/src/app/pages/projects/project-details/settings/project-information/project-information.component.html
@@ -1,18 +1,0 @@
-<div>
-  <h4>Project Settings</h4>
-
-  <mat-list>
-    <mat-list-item>
-      <span matListItemTitle>Id</span>
-      <span matListItemLine>{{ project().id }}</span>
-    </mat-list-item>
-    <mat-list-item>
-      <span matListItemTitle>Name</span>
-      <span>{{ project().id }}</span>
-    </mat-list-item>
-    <mat-list-item>
-      <span matListItemTitle>Creation Date</span>
-      <span>{{ project().creationDate | date }}</span>
-    </mat-list-item>
-  </mat-list>
-</div>
